shorter experience to generate form
you can now go for "rails g form foo_bar" instead of "rails g
hyperactiveform:form foo_bar"
